+++ driver-core/src/main/java/com/datastax/driver/core/ColumnMetadata.java
@@ -47,10 +47,7 @@ public class ColumnMetadata {
             String name = row.getString(COLUMN_NAME);
 
             String validator = row.getString(VALIDATOR);
-            // Ugly special case for TimestampType as we don't have it yet. We should get rid of that later.
-            if (validator.equals("org.apache.cassandra.db.marshal.TimestampType"))
-                validator = "org.apache.cassandra.db.marshal.DateType";
-            AbstractType<?> t = TypeParser.parse(validator);
+            AbstractType<?> t = TypeParser.parse(TableMetadata.fixTimestampType(validator));
             ColumnMetadata cm = new ColumnMetadata(tm, name, Codec.rawTypeToDataType(t), row);
             tm.add(cm);
             return cm;