+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/Delete.java
@@ -50,7 +50,7 @@ public class Delete extends BuiltStatement {
     }
 
     @Override
-    protected String buildQueryString() {
+    protected StringBuilder buildQueryString() {
         StringBuilder builder = new StringBuilder();
 
         builder.append("DELETE ");
@@ -71,7 +71,7 @@ public class Delete extends BuiltStatement {
             Utils.joinAndAppend(builder, " AND ", where.clauses);
         }
 
-        return builder.toString();
+        return builder;
     }
 
     /**